[POC] Get Control keys working under macOS
This was tested commingled with some reversions to the
Vt102Emulation::remapKeyModifiersForMac() method, so I don't know for a
fact that this change stands on its own.

I think it probably does, though - the layered tangle of Control key
options and trickery doesn't make any sense to me at all in the presence
of the AA_MacDontSwapCtrlAndMeta option in the main app's codebase, as
noted in the change.